MFRC522: Move ReadUID example to example_test.go
