+++ experimental/devices/mfrc522/example_test.go
@@ -102,3 +102,75 @@ func Example() {
 		}
 	}
 }
+
+func ExampleDev_ReadUID() {
+	// Make sure periph is initialized.
+	if _, err := host.Init(); err != nil {
+		log.Fatal(err)
+	}
+
+	// Using SPI as an example. See package ./spi/spireg for more details.
+	p, err := spireg.Open("")
+	if err != nil {
+		log.Fatal(err)
+	}
+	defer p.Close()
+
+	rfid, err := mfrc522.NewSPI(p, rpi.P1_22, rpi.P1_18)
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	// Idling device on exit.
+	defer rfid.Halt()
+
+	// Setting the antenna signal strength.
+	rfid.SetAntennaGain(5)
+
+	timedOut := false
+	cb := make(chan []byte)
+	timer := time.NewTimer(10 * time.Second)
+
+	// Stopping timer, flagging reader thread as timed out
+	defer func() {
+		timer.Stop()
+		timedOut = true
+		close(cb)
+	}()
+
+	go func() {
+		log.Printf("Started %s", rfid.String())
+
+		for {
+			// Trying to read card UID.
+			uid, err := rfid.ReadUID(10 * time.Second)
+
+			// If main thread timed out just exiting.
+			if timedOut {
+				return
+			}
+
+			// Some devices tend to send wrong data while RFID chip is already detected
+			// but still "too far" from a receiver.
+			// Especially some cheap CN clones which you can find on GearBest, AliExpress, etc.
+			// This will suppress such errors.
+			if err != nil {
+				continue
+			}
+
+			cb <- uid
+			return
+		}
+	}()
+
+	for {
+		select {
+		case <-timer.C:
+			log.Fatal("Didn't receive device data")
+			return
+		case data := <-cb:
+			log.Println("UID:", hex.EncodeToString(data))
+			return
+		}
+	}
+}
